+++ __test__/holiday/holiday.test.ts
@@ -3,7 +3,7 @@ import { describe, expect, it } from "vitest";
 import { isHoliday } from "../../src";
 import type { Holiday } from "../../src/interfaces";
 
-describe("Holiday Validation", () => {
+describe("Holiday validation", () => {
   it("should return false when given null as BasicJewishDate and holidayList is empty", () => {
     const result = isHoliday(null as unknown as BasicJewishDate, []);
     expect(result).toBe(false);