+++ src/yomTov/index.ts
@@ -1,1 +1,2 @@
 export * from "./yomTov";
+export * from "./erevYomTov";