+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 1.1.0
+
+- Added the following methods:
+  - `isErevShabbat` - Determines whether a given date is Erev Shabbat (the day before Shabbat).
+  - `isErevYomTov` - Determines whether a given date is a Erev Yom Tov (the day before a Jewish holiday).
+  - `isCholHaMoed` - Determines whether a given date falls within Chol HaMoed.
+  - `isChanukah` - Determines whether a given date falls during Chanukah.
+- Change typos in tests
+- Upgrade all build tools to latest.
+
+
 ## 1.0.3
 
 - Fix issue in isYomTov not returning true for Simchat Torah.